requirements.txt
discord.py
python-dotenv
langchain
openai
unstructured
promptlayer